Move CLConnectionManager from generic util folder to engine API RPC folder
