+++ packages/client/src/rpc/modules/engine/index.ts
@@ -1,3 +1,4 @@
+export * from './CLConnectionManager'
 export * from './engine'
 export * from './types'
 export * from './util'
